+++ databasez/__init__.py
@@ -1,5 +1,5 @@
 from databasez.core import Database, DatabaseURL
 
-__version__ = "0.9.3"
+__version__ = "0.9.4"
 
 __all__ = ["Database", "DatabaseURL"]